Enhanced the file name detection.
This is a slight improvement, to ignore modules in /private. This
works on my machine but ymmv, e.g. they might be in /tmp. To make this
work with compilation-mode, I've used a horrendous regexp hack. A
cleaner approach would be to enhance compilation-mode to have a better
way of "setting" the filename, or the regexp could be cleaned up if
Emacs supported lookahead assertions.

Someone made a patch for lookahead assertions, but it hasn't been
accepted: <https://debbugs.gnu.org/cgi/bugreport.cgi?bug=5393>. This
no longer applies cleanly, but there is an up-to-date patch on
github. This note here is a reminder for why the regexp is so awful.
This place is not a place of honor.