Add indentation for TLA+ mode
For now, under the hood, simply re-using the indentation for PlusCal
mode.  This seems to actually work fairly well -- so many constructs mix
between the two languages, this has been quire workable in practice.

Left in place a few distinct functions at the top-level, to enable
straightforward splitting of functionality in the future, if desired.