Add basic support for aligning /\, \/, etc
For now, very simple rule, focused on not breaking semantics of
clauses (where whitespace indentation is signficant):

 - Only apply the rule if line starts with /\, \/, etc

 - If the line above also starts with one of those symbols...

 - Indent to start of that line...

 - *Unless* the current line is already further indented, in which case,
    do nothing.

This ensures that if, e.g. the programmer has typed:

 /\
    \/ x = 1
    \/ y = 2
 /\
 ...

Then hitting tab *won't* pull one of the \/'s to the left, which would
change the meaning.

It's still a bit clunky -- e.g. for reasons that are unclear to me right
now, the point doesn't go to the start of the line, even when the line
gets properly indented. Also, ideally, I'd like to set it up to cycle
through indent lines on repeated hits to tab, but not there yet.

But this has helped making writing TLA+ workable, so felt ready to share.